Support custom Sprockets manifest filenames
In Rails 4, `config.assets.manifest` must include the manifest filename,
eg.

    config.assets.manifest = "public/assets/manifest-foo.json"

This change adds support for such configurations.

Note that while it's invalid for the manifest option to be a directory
in Rails 4, we are more forgiving and also accept manifest paths
paths that don't contain the filename, in order to
preserve backwards compatibility with previous requirejs-rails versions
and Rails 3 apps.